fix: coerce output_shape list to first shape (see commit details)
Input layers seem to have a list of output shapes. For example:

```python
>>> input_layer.output_shape
[(None, 224, 224, 3)]

>>> regular_layer.output_shape
(None, 224, 224, 3)
```

As a workaround, to coerce it to a single tuple, we check if the
`output_shape` is a `list` and take the first element if it is:

```python
output_shape = layer.output_shape
if isinstance(output_shape, list):
    output_shape = output_shape[0]
```

Furthermore, if there is more than one item in the list, raise an
exception.